move FloatingPoint to util from rlp.util
